Prevent float64 truncation of integer larger than uint32

+++ src/cpp/type-conversion.cpp
@@ -59,7 +59,9 @@ namespace TypeConversion {
       }
       case G_TYPE_UINT64: {
         if (js_value.IsNumber()) {
-          g_value_set_uint64(out_value, js_value.As<Napi::Number>().Uint32Value());
+          g_value_set_uint64(
+            out_value, static_cast<uint64_t>(js_value.As<Napi::Number>().DoubleValue())
+          );
           return true;
         } else if (js_value.IsBigInt()) {
           bool lossless;
@@ -118,11 +120,11 @@ namespace TypeConversion {
             std::string flags_str = js_value.As<Napi::String>().Utf8Value();
             GFlagsClass *flags_class = G_FLAGS_CLASS(g_type_class_ref(target_type));
             guint flags_value = 0;
-            
+
             // Try to parse the string as flag names
             gchar **flag_names = g_strsplit(flags_str.c_str(), "+", -1);
             gboolean success = TRUE;
-            
+
             for (gint i = 0; flag_names[i] != NULL; i++) {
               g_strstrip(flag_names[i]); // Remove whitespace
               GFlagsValue *flag_val = g_flags_get_value_by_nick(flags_class, flag_names[i]);
@@ -136,10 +138,10 @@ namespace TypeConversion {
                 break;
               }
             }
-            
+
             g_strfreev(flag_names);
             g_type_class_unref(flags_class);
-            
+
             if (success) {
               g_value_set_flags(out_value, flags_value);
               return true;
@@ -239,7 +241,7 @@ namespace TypeConversion {
     } else if (G_TYPE_IS_FLAGS(G_VALUE_TYPE(gvalue))) {
       guint flags_val = g_value_get_flags(gvalue);
       GFlagsClass *flags_class = G_FLAGS_CLASS(g_type_class_ref(G_VALUE_TYPE(gvalue)));
-      
+
       // Manual flags to string conversion
       std::string flags_str = "";
       if (flags_val == 0) {
@@ -261,7 +263,7 @@ namespace TypeConversion {
           }
         }
       }
-      
+
       if (!flags_str.empty()) {
         Napi::Value result = Napi::String::New(env, flags_str);
         g_type_class_unref(flags_class);
@@ -312,10 +314,10 @@ namespace TypeConversion {
     if (js_value.IsNull()) {
       return env.Null();
     }
-    
+
     // Create the standardized result object
     Napi::Object result = Napi::Object::New(env);
-    
+
     // Determine the type and set both type and value
     if (js_value.IsString() || js_value.IsNumber() || js_value.IsBoolean()) {
       result.Set("type", Napi::String::New(env, "primitive"));
@@ -337,7 +339,7 @@ namespace TypeConversion {
       // Default to primitive for unknown types
       result.Set("type", Napi::String::New(env, "primitive"));
     }
-    
+
     result.Set("value", js_value);
     return result;
   }
